ELEC-486: Autostart (#22)
Fixes a bug in the autostart script and makes the default values in the display all 0s.

To update the install on the PI:

```bash
cd /home/pi/telemetry
rm -rf .
```

Download the release `arm32_linux.zip` from the releases page.

Unzip the contents of the zipfile in `/home/pi/telemetry`. I have already configured the PI to autostart the rest.